+++ include/eosio.wps/eosio.wps.hpp
@@ -95,10 +95,9 @@ typedef eosio::multi_index< "drafts"_n, drafts_row> drafts_table;
  * - `{string} title` - proposal title
  * - `{asset} budget` - monthly budget payment request
  * - `{uint8_t} duration` - monthly budget duration (maximum of 6 months)
- * - `{name} status` - current status of proposal (draft/active/completed/expired)
- * - `{asset} deposit` - deposit required to active proposal
- * - `{asset} payments` - payments made to WPS proposal
  * - `{map<name, string>} proposal_json` - a sorted container of <key, value>
+ * - `{time_point_sec} start` - start of voting period (UTC)
+ * - `{time_point_sec} end` - end of voting period (UTC)
  *
  * ### example
  *
@@ -109,13 +108,12 @@ typedef eosio::multi_index< "drafts"_n, drafts_row> drafts_table;
  *   "title": "My WPS",
  *   "budget": "500.0000 EOS",
  *   "duration": 1,
- *   "status": "draft",
- *   "deposit": "0.0000 EOS",
- *   "payments": "0.0000 EOS",
  *   "proposal_json": [
  *     { "key": "category", "value": "other" },
  *     { "key": "region", "value": "global" }
- *   ]
+ *   ],
+ *   "start": "2019-11-05T12:10:00",
+ *   "end": "2019-12-01T00:00:00"
  * }
  * ```
  */
@@ -125,28 +123,20 @@ struct [[eosio::table("proposals"), eosio::contract("eosio.wps")]] proposals_row
     eosio::string                           title;
     eosio::asset                            budget;
     uint8_t                                 duration;
-    eosio::name                             status;
-    eosio::asset                            deposit;
-    eosio::asset                            payments;
     std::map<eosio::name, eosio::string>    proposal_json;
+    eosio::time_point_sec                   start;
+    eosio::time_point_sec                   end;
 
     uint64_t primary_key() const { return proposal_name.value; }
-    uint64_t by_status() const { return status.value; }
 };
 
-typedef eosio::multi_index< "proposals"_n, proposals_row,
-    indexed_by<"bystatus"_n, const_mem_fun<proposals_row, uint64_t, &proposals_row::by_status>>
-> proposals_table;
+typedef eosio::multi_index< "proposals"_n, proposals_row> proposals_table;
 
 /**
  * ## TABLE `votes`
  *
  * - `{name} proposal_name` - The proposal's name, its ID among all proposals
- * - `{name} status` - current status of proposal (draft/active/completed/expired)
  * - `{int16_t} total_net_votes` - total net votes
- * - `{time_point_sec} voting_period` - active voting period (UTC)
- * - `{time_point_sec} start` - start of voting period (UTC)
- * - `{time_point_sec} end` - end of voting period (UTC)
  * - `{map<name, name>} votes` - a sorted container of <voter, vote>
  *
  * ### example
@@ -154,11 +144,7 @@ typedef eosio::multi_index< "proposals"_n, proposals_row,
  * ```json
  * {
  *   "proposal_name": "mywps",
- *   "status": "active",
  *   "total_net_votes": 2,
- *   "voting_period": "2019-11-01T00:00:00",
- *   "start": "2019-11-05T12:10:00",
- *   "end": "2019-12-01T00:00:00",
  *   "votes": [
  *      { "key": "mybp1", "value": "yes" },
  *      { "key": "mybp2", "value": "no" },
@@ -171,22 +157,13 @@ typedef eosio::multi_index< "proposals"_n, proposals_row,
  */
 struct [[eosio::table("votes"), eosio::contract("eosio.wps")]] votes_row {
     eosio::name                         proposal_name;
-    eosio::name                         status;
     int16_t                             total_net_votes;
-    eosio::time_point_sec               voting_period;
-    eosio::time_point_sec               start;
-    eosio::time_point_sec               end;
     std::map<eosio::name, eosio::name>  votes;
 
     uint64_t primary_key() const { return proposal_name.value; }
-    uint64_t by_status() const { return status.value; }
-    uint64_t by_voting_period() const { return voting_period.sec_since_epoch(); }
 };
 
-typedef eosio::multi_index< "votes"_n, votes_row,
-    indexed_by<"bystatus"_n, const_mem_fun<votes_row, uint64_t, &votes_row::by_status>>,
-    indexed_by<"byperiod"_n, const_mem_fun<votes_row, uint64_t, &votes_row::by_voting_period>>
-> votes_table;
+typedef eosio::multi_index< "votes"_n, votes_row> votes_table;
 
 /**
  * ## TABLE `state`